+++ ray.cu
@@ -52,8 +52,8 @@ __host__ __device__ void cross(const ray l_1, const line l_2, float& t_1, float&
 	cross(line(l_1), line(l_2), t_1, t_2);
 	if (0 > t_1)
 	{
-		t_1 = DBL_MAX;
-		t_2 = DBL_MAX;
+		t_1 = FLT_MAX;
+		t_2 = FLT_MAX;
 	}
 }
 
@@ -62,8 +62,8 @@ __host__ __device__ void cross(const ray l_1, const ray l_2, float& t_1, float& 
 	cross(line(l_1), line(l_2), t_1, t_2);
 	if ((0 > t_1) || (0 > t_2))
 	{
-		t_1 = DBL_MAX;
-		t_2 = DBL_MAX;
+		t_1 = FLT_MAX;
+		t_2 = FLT_MAX;
 	}
 }
 
@@ -72,8 +72,8 @@ __host__ __device__ void cross(const ray l_1, const seg l_2, float& t_1, float& 
 	cross(line(l_1), line(l_2), t_1, t_2);
 	if ((0 > t_1) || (0 > t_2) || (t_2 > l_2.dist))
 	{
-		t_1 = DBL_MAX;
-		t_2 = DBL_MAX;
+		t_1 = FLT_MAX;
+		t_2 = FLT_MAX;
 	}
 }
 
@@ -81,13 +81,13 @@ __host__ __device__ point cross(const ray l_1, const line l_2)
 {
 	float t_1, t_2;
 	cross(l_1, l_2, t_1, t_2);
-	if (t_1 != DBL_MAX)
+	if (t_1 != FLT_MAX)
 	{
 		return l_1.point_get(t_1);
 	}
 	else
 	{
-		return point(DBL_MAX, DBL_MAX);
+		return point(FLT_MAX, FLT_MAX);
 	}
 }
 
@@ -95,13 +95,13 @@ __host__ __device__ point cross(const ray l_1, const ray l_2)
 {
 	float t_1, t_2;
 	cross(l_1, l_2, t_1, t_2);
-	if (t_1 != DBL_MAX)
+	if (t_1 != FLT_MAX)
 	{
 		return l_1.point_get(t_1);
 	}
 	else
 	{
-		return point(DBL_MAX, DBL_MAX);
+		return point(FLT_MAX, FLT_MAX);
 	}
 }
 
@@ -109,13 +109,13 @@ __host__ __device__ point cross(const ray l_1, const seg l_2)
 {
 	float t_1, t_2;
 	cross(l_1, l_2, t_1, t_2);
-	if (t_1 != DBL_MAX)
+	if (t_1 != FLT_MAX)
 	{
 		return l_1.point_get(t_1);
 	}
 	else
 	{
-		return point(DBL_MAX, DBL_MAX);
+		return point(FLT_MAX, FLT_MAX);
 	}
 }
 
@@ -123,7 +123,7 @@ __host__ __device__ bool is_cross(const ray l_1, const seg l_2)
 {
 	float t_1, t_2;
 	cross(l_1, l_2, t_1, t_2);
-	if (t_1 == DBL_MAX)
+	if (t_1 == FLT_MAX)
 	{
 		return false;
 	}
@@ -134,7 +134,7 @@ __host__ __device__ bool is_cross(const ray l_1, const line l_2)
 {
 	float t_1, t_2;
 	cross(l_1, l_2, t_1, t_2);
-	if (t_1 == DBL_MAX)
+	if (t_1 == FLT_MAX)
 	{
 		return false;
 	}
@@ -145,7 +145,7 @@ __host__ __device__ bool is_cross(const ray l_1, const ray l_2)
 {
 	float t_1, t_2;
 	cross(l_1, l_2, t_1, t_2);
-	if (t_1 == DBL_MAX)
+	if (t_1 == FLT_MAX)
 	{
 		return false;
 	}
